+++ R/margot_plot_hist.R
@@ -32,11 +32,13 @@
 #' @return A ggplot2 object representing the colored histogram.
 #'
 #' @examples
+#' \dontrun{
 #' df <- data.frame(value = rnorm(1000))
 #' coloured_histogram_quantiles(df, "value", n_divisions = 4)
 #'
 #' # With custom breaks
 #' coloured_histogram_quantiles(df, "value", custom_breaks = c(-2, -1, 0, 1, 2))
+#' }
 #'
 #' @import ggplot2
 #' @importFrom ggokabeito palette_okabe_ito