+++ R/margot_prop_missing.R
@@ -12,11 +12,11 @@
 #'
 #' @examples
 #' # Example using a dataset with a wave column
-#' # assume dat_long has a column called wave
+#' dat_long <- data.frame(wave = c(0, 0, 1), outcome = c(1, NA, 3))
 #' margot_prop_missing(dat_long)
 #'
 #' # Example using a dataset without a wave column
-#' # assume some_data is a dataset with no wave column
+#' some_data <- data.frame(outcome = c(1, NA, 3))
 #' margot_prop_missing(some_data)
 margot_prop_missing <- function(data, wave_col = "wave") {
   if (!wave_col %in% names(data)) {